requirements.txt
flask==2.3.3
flask-socketio==5.3.6
python-dotenv==1.0.0
pymongo==4.6.1
paho-mqtt==1.6.1
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.6.0
joblib==1.3.2
gunicorn==21.2.0
eventlet==0.33.3
dnspython==2.4.2 